using size_t across all containers to represent their capacity
NOTE: the public API has changed since now all positions and indices are
      expected to be size_t instead of uint32_t

+++ src/binheap.h
@@ -258,7 +258,7 @@ int binheap_walk(binheap_t *bh, binheap_walk_callback_t cb, void *priv);
  * @param bh A valid pointer to an initialized binheap_t structure
  * @return the actual number of items in the heap
  */
-uint32_t binheap_count(binheap_t *bh);
+size_t binheap_count(binheap_t *bh);
 
 /**
  * @brief Default callbacks to handle 16bit signed integer keys
